+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.0 (WIP)
 - TWEAK: Renamed module and database to `hyperglot`
 - TWEAK: `--strict-support` refactored to `--validity` with default `weak` to pick the level of required validity on the languages that should get matched
+- TWEAK: Saving and validating enforces removal of superflous mark characters that are getting implicitly extracted via glyph decomposition
 
 ## 0.1.14 (27.5.2020)
 - FIX: Refined character parsing to also include the encoded form of any decomposable glyphs